perf(parser): reject autolink candidates on `://` instead of `:` (-4% parse)
The GFM autolink post-pass costs 10-14% of a parse on the bundled corpora
— ablating it entirely takes rust-book from 2.88 ms to 2.51 ms. Almost
none of that is spent rewriting anything; it is the cost of deciding
there is nothing to rewrite.

Two things were paying for that decision.

The block-level gate asked for a bare `:`. Documentation prose is full of
`Note:`, `1:1` and `see Chapter 3:`, so the gate passed constantly and
dragged the block through the whole node-tree walk, the text coalescing,
and a per-node substring search. But a scheme candidate needs the whole
`://`, and asking for that is the same one-pass `memmem` — it just
rejects far more. The gate now probes `@`/`&` with one `memchr2`, then
`www.`, then `://`, ordered by how cheaply each rejects.

Per text node, `find_candidate` ran up to six whole-node scans: one
`memchr2`, four `memmem` searchers, and an `@` walk. `http://`,
`https://` and `ftp://` all end in `://` and differ only in the name in
front, so one pass for `://` finds every scheme candidate in position
order and the name is recovered by looking back from the match — three
searchers become one. And `www.` is the only needle a node can hold
without a `:` or `@` in it, so whether it is worth searching for is a
property of the block, not the node: the gate now hands that verdict
down and ordinary prose leaves `find_candidate` after a single
vectorized byte scan with no substring searcher touched.

Parse, best of 60 iterations:

  rust-book            2.884 -> 2.718 ms  -5.8%
  typescript-handbook  3.882 -> 3.757 ms  -3.2%
  vite-docs            1.272 -> 1.226 ms  -3.6%
  vue-docs             2.234 -> 2.150 ms  -3.8%

Parse + render:

  rust-book            4.131 -> 3.963 ms  -4.1%
  typescript-handbook  5.938 -> 5.828 ms  -1.9%
  vite-docs            1.839 -> 1.778 ms  -3.3%
  vue-docs             3.347 -> 3.258 ms  -2.7%

Both gates stay conservative supersets. `:`, `@` and `.` are never
inline-special, so a needle that survives into a text node was spelled
verbatim in the block's raw source; the one indirection is entities,
which `&` covers by keeping the pass — and the `www.` search — on.

The scans move to a `gfm_autolink/scan.rs` submodule to keep both files
inside the 350-line source cap.
